move interface from service to common folder
